requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic>=1.10.0,<3.0.0
pydantic-settings>=2.0.0
watchdog==3.0.0
python-dotenv==1.0.0
websockets==11.0.3
sqlalchemy==2.0.19
alembic==1.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.7
requests==2.31.0
aiofiles==23.1.0
aiohttp==3.8.5
email-validator>=2.0.0
